+++ apps/web/src/app/admin-console/organizations/shared/components/access-selector/access-selector.stories.ts
@@ -1,7 +1,7 @@
 import { importProvidersFrom } from "@angular/core";
 import { FormBuilder, FormsModule, ReactiveFormsModule } from "@angular/forms";
-import { action } from "@storybook/addon-actions";
 import { applicationConfig, Meta, moduleMetadata, StoryObj } from "@storybook/angular";
+
 
 import { JslibModule } from "@bitwarden/angular/jslib.module";
 import {
@@ -22,7 +22,8 @@ import {
 import { PreloadedEnglishI18nModule } from "../../../../../core/tests";
 
 import { AccessSelectorComponent } from "./access-selector.component";
-import { AccessItemType, AccessItemView, CollectionPermission } from "./access-selector.models";
+import { AccessItemType, CollectionPermission } from "./access-selector.models";
+import { actionsData, itemsFactory } from "./storybook-utils";
 import { UserTypePipe } from "./user-type.pipe";
 
 export default {
@@ -58,47 +59,6 @@ export default {
 // TODO: This is a workaround since this story does weird things.
 type Story = StoryObj<any>;
 
-const actionsData = {
-  onValueChanged: action("onValueChanged"),
-  onSubmit: action("onSubmit"),
-};
-
-/**
- * Factory to help build semi-realistic looking items
- * @param n - The number of items to build
- * @param type - Which type to build
- */
-const itemsFactory = (n: number, type: AccessItemType) => {
-  return [...Array(n)].map((_: unknown, id: number) => {
-    const item: AccessItemView = {
-      id: id.toString(),
-      type: type,
-    } as AccessItemView;
-
-    switch (item.type) {
-      case AccessItemType.Collection:
-        item.labelName = item.listName = `Collection ${id}`;
-        item.id = item.id + "c";
-        item.parentGrouping = "Collection Parent Group " + ((id % 2) + 1);
-        break;
-      case AccessItemType.Group:
-        item.labelName = item.listName = `Group ${id}`;
-        item.id = item.id + "g";
-        break;
-      case AccessItemType.Member:
-        item.id = item.id + "m";
-        item.email = `member${id}@email.com`;
-        item.status = id % 3 == 0 ? 0 : 2;
-        item.labelName = item.status == 2 ? `Member ${id}` : item.email;
-        item.listName = item.status == 2 ? `${item.labelName} (${item.email})` : item.email;
-        item.role = id % 5;
-        break;
-    }
-
-    return item;
-  });
-};
-
 const sampleMembers = itemsFactory(10, AccessItemType.Member);
 const sampleGroups = itemsFactory(6, AccessItemType.Group);
 
@@ -127,48 +87,6 @@ const StandaloneAccessSelectorRender = (args: any) => ({
   `,
 });
 
-const DialogAccessSelectorRender = (args: any) => ({
-  props: {
-    items: [],
-    valueChanged: actionsData.onValueChanged,
-    initialValue: [],
-    ...args,
-  },
-  template: `
-    <bit-dialog [dialogSize]="dialogSize" [disablePadding]="disablePadding">
-      <span bitDialogTitle>Access selector</span>
-      <span bitDialogContent>
-        <bit-access-selector
-          (ngModelChange)="valueChanged($event)"
-          [ngModel]="initialValue"
-          [items]="items"
-          [disabled]="disabled"
-          [columnHeader]="columnHeader"
-          [showGroupColumn]="showGroupColumn"
-          [selectorLabelText]="selectorLabelText"
-          [selectorHelpText]="selectorHelpText"
-          [emptySelectionText]="emptySelectionText"
-          [permissionMode]="permissionMode"
-          [showMemberRoles]="showMemberRoles"
-        ></bit-access-selector>
-      </span>
-      <ng-container bitDialogFooter>
-        <button bitButton buttonType="primary">Save</button>
-        <button bitButton buttonType="secondary">Cancel</button>
-        <button
-          class="tw-ml-auto"
-          bitIconButton="bwi-trash"
-          buttonType="danger"
-          size="default"
-          title="Delete"
-          aria-label="Delete"></button>
-      </ng-container>
-    </bit-dialog>
-  `,
-});
-
-const dialogAccessItems = itemsFactory(10, AccessItemType.Collection);
-
 const memberCollectionAccessItems = itemsFactory(3, AccessItemType.Collection).concat([
   {
     id: "c1-group1",
@@ -189,22 +107,6 @@ const memberCollectionAccessItems = itemsFactory(3, AccessItemType.Collection).c
     readonly: true,
   },
 ]);
-
-export const Dialog: Story = {
-  args: {
-    permissionMode: "edit",
-    showMemberRoles: false,
-    showGroupColumn: true,
-    columnHeader: "Collection",
-    selectorLabelText: "Select Collections",
-    selectorHelpText: "Some helper text describing what this does",
-    emptySelectionText: "No collections added",
-    disabled: false,
-    initialValue: [] as any[],
-    items: dialogAccessItems,
-  },
-  render: DialogAccessSelectorRender,
-};
 
 export const MemberCollectionAccess: Story = {
   args: {
